tertiary biopores flux
git-svn-id: https://daisy-model.googlecode.com/svn/trunk@3613 5f5bf4bb-3343-0410-8278-4fa8eb3319dc

+++ lexer_flux.h
@@ -31,7 +31,7 @@ class LexerFlux : public LexerTable
 private:
   symbol array_name;
   symbol array_dimension;
-  std::vector<double> array_c;
+  std::vector<int> array_c;
   std::vector<double> center_z;
   std::vector<double> center_x;
   std::vector<int> flux_from;
